String Methods - JavaScript String repeat()
The repeat() method returns a string with a number of copies of a string.

The repeat() method returns a new string.

The repeat() method does not change the original string.

repeat() returns a new string with a number of copies of a string:

let text = "Hello world!";
let result = text.repeat(2);
Hello world!Hello world!