this Keyword - Explicit Function Binding - call method
The call() and apply() methods are predefined JavaScript methods.

They can both be used to call an object method with another object as argument.

See Also:
The Function call() Method

The Function apply() Method

The Function bind() Method

The example below calls person1.fullName with person2 as an argument, this refers to person2, even if fullName is a method of person1:

Example

const person1 = {
  fullName: function() {
    return this.firstName + " " + this.lastName;
  }
}

const person2 = {
  firstName:"John",
  lastName: "Doe",
}

// Return "John Doe":
person1.fullName.call(person2);