Mistakes - Undefined is Not Null
JavaScript objects, variables, properties, and methods can be undefined.

In addition, empty JavaScript objects can have the value null.

This can make it a little bit difficult to test if an object is empty.

You can test if an object exists by testing if the type is undefined:

Example:
if (typeof myObj === "undefined")
// Result :true