Mistakes - Undefined is Not Null
JavaScript objects, variables, properties, and methods can be undefined.

In addition, empty JavaScript objects can have the value null.

This can make it a little bit difficult to test if an object is empty.

You can test if an object exists by testing if the type is undefined:

But you cannot test if an object is null, because this will throw an error if the object is undefined:

Incorrect:
if (myObj === null)
To solve this problem, you must test if an object is not null, and not undefined.

But this can still throw an error:

Incorrect:
if (myObj !== null && typeof myObj !== "undefined")
Because of this, you must test for not undefined before you can test for not null:

Correct:
if (typeof myObj !== "undefined" && myObj !== null)