Browser BOM Navigator - Browser Application Code Name
The appCodeName property returns the application code name of the browser

navigator.appCodeName

The appCodeName property returns the code name of the browser.

Do not rely on it! "Mozilla" is the application code name for Chrome, Firefox, IE, Safari, and Opera.

Warning

This property is removed (deprecated) in the latest web standard.

Most browsers (IE11, Chrome, Firefox, Safari, Opera) returns Mozilla as appCodeName.